Move more code from answer_queries to notification
